correcao

+++ app/Http/Controllers/Event/Meeting/MeetingController.php
@@ -65,6 +65,7 @@ class MeetingController extends Controller
      */
     public function show($id)
     {
+        $this->authorize('show',meeting::class);
         //verifica se a origem é de notificaçoes
         if (request()->notification) {
             $notification = Notification::find(request()->notification);
@@ -108,6 +109,7 @@ class MeetingController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('show',meeting::class);
         //verifica se a origem é de notificaçoes
         if (request()->notification) {
             $notification = Notification::find(request()->notification);
@@ -189,6 +191,7 @@ class MeetingController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete',meeting::class);
         $afectedRows = $this->service->destroy($id);
         if ($afectedRows) {
             return redirect()->route('meeting.list');
